requirements.txt
Django==3.0.7
python-dotenv==0.13.0
Pillow==7.1.2
django-admin-sortable2==0.7.6
